mod: restructured pages, added page title and mother-layout
