+++ app/(dashboard)/dashboard/layout.tsx
@@ -1,0 +1,16 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+    title: "Dashboard",
+    description: "EscrowNet Dashboard page",
+};
+
+export default function DashboardLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <>{children}</>
+    );
+}
