+++ app/(dashboard)/escrow/[id]/page.tsx
@@ -1,5 +1,5 @@
 import DashboardContentContainer from "@/components/DashboardContentContainer";
-import Header from "../../../components/dashboardHeader";
+import Header from "../../../../components/dashboardHeader";
 
 const SignEscrow = () => {
   return (
