--- app/dashboard/layout.tsx
@@ -1,32 +1,0 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "../globals.css";
-import Sidebar from "@/components/Sidebar";
-
-const geistSans = Geist({
-    variable: "--font-geist-sans",
-    subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-    variable: "--font-geist-mono",
-    subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-    title: "Create Next App",
-    description: "Generated by create next app",
-};
-
-export default function RootLayout({
-    children,
-}: Readonly<{
-    children: React.ReactNode;
-}>) {
-    return (
-        <div className="flex">
-            <Sidebar />
-            <main className="ml-64 flex-1 ">{children}</main>
-        </div>
-    );
-}
